Fix <input> alignment issue.
typeahead messes up the alignment of the input. There was already a fix
in place for it, but I believe it was only tested on Windows
not exactly correct on OS X.

I'm not sure but I believe the problem is related to fonts rendering
differently between operating systems.

If I'm correct, this fix should work better since it uses relative units
instead of absolute units. In any case it *looks* better on OS X, though
we definitely need to test on Windows since that is where the majority
of users are.