Add task for building custom version of jQuery
We use a custom build of jQuery that omits the modules that aren't
absolutely necessary for the snippet & typeahead. This commit adds a
make task that makes a shallow clone of the jQuery repo, checks out the
tag for the version we are using (v2.1.1) and runs the grunt task to
create our custom build.

After making the build, `sed` is used to cut off the last line before
copying it to the working directory. This is important for when we go to
create the bundle: the jQuery build process puts a source map
declaration at the end, which is a newline that starts with `//#`. When
we go to concatenate the files, this can mess things up.

This commit also does general cleanup and refactoring of the Makefile.