Include the snippet HTML in the index.
This is part of starting to test the snippet as it will be implemented
on about:home. Learning a little more how things are implemented, it
probably makes sense to put the HTML directly in <div id="snippets">
rather than loading it from JavaScript.

A counter-argument to this would be that if for some reason the
JavaScript fails, it'd be better to show no snippet at all versus one
that doesn't work when the user tries to interact with it.

I'll have to consider this and maybe get a second opinion, but for now
I'm going to include the HTML directly on the page.