Use data URI for the webmaker logo.
As part of the snippet guidelines all images must be in the form of data
URIs. This commit replaces the external src with a data URI src in the
snippet HTML. I've also created a `make` task for generating a data URI
for the logo image.

We are using a PNG asset currently (~150x150), which isn't ideal. I
found an EPS version and got someone with Illustrator to export it as an
SVG, but that didn't end up working so well: either because of the
original file, or the specific version of Illustrator that was used, the
SVG ended up containing a *lot* of bitmap data and was over 3 times as
big as the original PNG.