Implement build system with gulp.
A requirement for the snippet is that everything it needs to be
displayed must be in one file: scripts and styles inline and images data
URI'd. I created a Makefile for building the script bundle and minifying
the CSS, but it became complicated when trying to pull all those pieces
together.

Makefiles are also not really cross operating system friendly
because A) Windows doesn't have `make`, B) OS X doesn't come with `make`
out of the box and C) even on POSIX systems the tools in the pipeline
can vary slightly.

I looked at some of the popular node.js based build systems and decided
to go with gulp, mainly because it was easier to understand upfront and
looked like it would be more ergonomic for me coming from `make` (code
based vs configuration based).

There are probably still some minor optimizations that can be made, but
overall I'm happy with the results.