Fix typeahead to trigger "change" event
When selecting a suggestion, typeahead was setting the value of the input
field without triggering a "change" event.

I fixed this in a local copy of an un-minified version typeahead and
added `typeahead` as a gulp task to build the minified version.

I also [submitted a pull request][1] upstream to typeahead for the
"change" event trigger fix.

[1]: https://github.com/twitter/typeahead.js/pull/911